fix: removed warnings

+++ src/paho_connection_manager.cpp
@@ -53,7 +53,7 @@ bool addConnection(PAHOMQTTConnection *connection) {
 
 bool removeConnection(PAHOMQTTConnection *connection) {
   std::unique_lock<std::mutex> lck(connectionMutex);
-  for (int i = 0; i < connections.size(); i++) {
+  for (size_t i = 0; i < connections.size(); i++) {
     if (connections[i] == connection) {
       connections.erase(connections.begin() + i);
       return true;
